move all colorscale helpers to ./helpers.js and ./scales.js
