--- src/components/colorscale/default_scale.js
@@ -1,14 +1,0 @@
-/**
-* Copyright 2012-2018, Plotly, Inc.
-* All rights reserved.
-*
-* This source code is licensed under the MIT license found in the
-* LICENSE file in the root directory of this source tree.
-*/
-
-'use strict';
-
-var scales = require('./scales');
-
-
-module.exports = scales.RdBu;
