+++ src/components/colorscale/index.js
@@ -6,35 +6,32 @@
 * LICENSE file in the root directory of this source tree.
 */
 
-
 'use strict';
 
-exports.moduleType = 'component';
+var scales = require('./scales');
+var helpers = require('./helpers');
 
-exports.name = 'colorscale';
+module.exports = {
+    moduleType: 'component',
+    name: 'colorscale',
 
-exports.scales = require('./scales');
+    attributes: require('./attributes'),
+    layoutAttributes: require('./layout_attributes'),
 
-exports.defaultScale = require('./default_scale');
+    supplyLayoutDefaults: require('./layout_defaults'),
+    handleDefaults: require('./defaults'),
 
-exports.attributes = require('./attributes');
+    calc: require('./calc'),
 
-exports.layoutAttributes = require('./layout_attributes');
+    // ./scales.js is required in lib/coerce.js ;
+    // it needs to be a seperate module to avoid circular a dependency
+    scales: scales.scales,
+    defaultScale: scales.defaultScale,
+    getScale: scales.get,
+    isValidScale: scales.isValid,
 
-exports.supplyLayoutDefaults = require('./layout_defaults');
-
-exports.handleDefaults = require('./defaults');
-
-exports.calc = require('./calc');
-
-exports.hasColorscale = require('./has_colorscale');
-
-exports.isValidScale = require('./is_valid_scale');
-
-exports.getScale = require('./get_scale');
-
-exports.flipScale = require('./flip_scale');
-
-exports.extractScale = require('./extract_scale');
-
-exports.makeColorScaleFunc = require('./make_color_scale_func');
+    hasColorscale: helpers.hasColorscale,
+    flipScale: helpers.flipScale,
+    extractScale: helpers.extractScale,
+    makeColorScaleFunc: helpers.makeColorScaleFunc
+};
